[EI-95] [api] rename: http-exception-filter 파일명 변경
